rename CivicProfileForms folders to ProfileForms
