+++ test/components/ProfileForms/BasicInfo.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { expect, it, describe } from 'vitest';
-import { BasicInfo } from '@components/CivicProfileForms';
+import { BasicInfo } from '@components/ProfileForms';
 
 describe('Basic Info Form', () => {
   it('renders', () => {
