+++ test/pages/PersonalProfile.test.jsx
@@ -3,7 +3,7 @@ import { render } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
 import { expect, it, describe } from 'vitest';
 import { Profile } from '@pages';
-import { PROFILE_FORM_LIST } from '@components/CivicProfileForms';
+import { PROFILE_FORM_LIST } from '@components/ProfileForms';
 
 describe('Profile', () => {
   const numLinks = PROFILE_FORM_LIST.length;
